Move conversion functions to new file.
